@@ -2,26 +2,36 @@ class Task {
   final String id;
   final String title;
   final DateTime dueDate;
+  final String priority;
+  final List<String>? tags;
+  final String description;
 
   Task({
     required this.id,
     required this.title,
     required this.dueDate,
+    this.priority = 'Medium',
+    this.tags,
+    this.description = '',
   });
 
-  factory Task.fromMap(Map<String, dynamic> map) {
+  Task copyWith({
+    String? id,
+    String? title,
+    DateTime? dueDate,
+    String? priority,
+    List<String>? tags,
+    String? description,
+  }) {
     return Task(
-      id: map['id'] ?? '',
-      title: map['title'] ?? '',
-      dueDate: DateTime.parse(map['dueDate']),
+      id: id ?? this.id,
+      title: title ?? this.title,
+      dueDate: dueDate ?? this.dueDate,
+      priority: priority ?? this.priority,
+      tags: tags ?? this.tags,
+      description: description ?? this.description,
     );
   }
 
-  Map<String, dynamic> toMap() {
-    return {
-      'id': id,
-      'title': title,
-      'dueDate': dueDate.toIso8601String(),
-    };
-  }
+  // Also update fromMap and toMap accordingly for persistence
 }
